s05/Lightning-template-Hydra-ci/pyproject.toml
[project]
name = "lightning-template-hydra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "coverage>=7.8.0",
    "datasets>=3.0.0",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "lightning[extra]>=2.5.1.post0",
    "loguru>=0.7.3",
    "pytest>=8.3.5",
    "rich>=13.9.4",
    "rootutils>=1.0.7",
    "tensorboard>=2.19.0",
    "timm>=1.0.15",
    "torchvision>=0.22.0",
]


[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == '__main__':",
    "pass",
    "raise ImportError",
    "if TYPE_CHECKING:"
]
ignore_errors = true

[tool.pytest.ini_options]
filterwarnings = [
    # Ignore Lightning warnings about num_workers during testing
    "ignore::UserWarning:lightning.pytorch.trainer.connectors.data_connector",
    # Ignore warning about self.log() without trainer reference
    "ignore:You are trying to `self.log\\(\\)` but the `self.trainer`:UserWarning",
]
